taskllogin

+++ app/Http/Controllers/TasksController.php
@@ -8,6 +8,7 @@ use App\Task;    // 追加
 
 class TasksController extends Controller
 {
+    
     /**
      * Display a listing of the resource.
      *
@@ -15,22 +16,23 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $data = [];
         if (\Auth::check()) { // 認証済みの場合
+        
             // 認証済みユーザを取得
             $user = \Auth::user();
-            
-            $data = [
-                'user' => $user,
-            ];
-        }
-        // メッセージ一覧を取得
-        $tasks = Task::all();
 
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            
+            // タスク一覧を取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            return view('tasks.index', [
+                'user' => $user,
+                'tasks' => $tasks,
+            ]);
+        }else{
+            return view('welcome');
+        }
+
     }
 
     /**
@@ -40,6 +42,7 @@ class TasksController extends Controller
      */
     public function create()
     {
+        
         $task = new Task;
 
         // メッセージ作成ビューを表示
@@ -56,6 +59,7 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
+     
       // バリデーション
         $this->validate($request, [
             'status' => 'required|max:10',
@@ -64,12 +68,14 @@ class TasksController extends Controller
         
         // メッセージを作成
         $task = new Task;
+        $task->user_id = \Auth::id();
         $task->status = $request->status;    // 追加
         $task->content = $request->content;
         $task->save();
 
         // トップページへリダイレクトさせる
         return redirect('/');
+       
     }
 
     /**
@@ -80,13 +86,17 @@ class TasksController extends Controller
      */
     public function show($id)
     {
+        
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-
+// 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を表示
+        if (\Auth::id() === $task->user_id) {
         // メッセージ詳細ビューでそれを表示
         return view('tasks.show', [
             'task' => $task,
-        ]);
+        ]);}else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -97,13 +107,17 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
+        
          // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-
+// 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、編集を表示
+        if (\Auth::id() === $task->user_id) {
         // メッセージ編集ビューでそれを表示
         return view('tasks.edit', [
             'task' => $task,
-        ]);
+        ]);}else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -115,6 +129,7 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
+       
         // バリデーション
         $this->validate($request, [
             'status' => 'required|max:10',
@@ -141,10 +156,14 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
+        
         // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        $task = \App\Task::findOrFail($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
